requirements.txt
Django~=4.2.7
psycopg2-binary
python-dateutil~=2.8.2
pytest~=7.4.3
requests~=2.31.0